Move focus font to styles.ts
