+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import "./globals.css";
 import { ReactNode } from "react";
-import { FOCUS_FONT } from "@/utils/consts";
+import { FOCUS_FONT } from "@/utils/styles";
 import Head from 'next/head';
 
 export const metadata: Metadata = {
