+++ src/utils/styles.ts
@@ -1,0 +1,6 @@
+import { Lato } from "next/font/google";
+
+export const FOCUS_FONT = Lato({
+  subsets: ['latin'],
+  weight: ['100', '300', '400', '700', '900'],
+});
